pyproject.toml
[project]
name = "nanobot-ai"
version = "0.1.3.post6"
description = "A lightweight personal AI assistant framework"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "nanobot contributors"}
]
keywords = ["ai", "agent", "chatbot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "websockets>=12.0",
    "websocket-client>=1.6.0",
    "httpx[socks]>=0.25.0",
    "loguru>=0.7.0",
    "readability-lxml>=0.8.0",
    "rich>=13.0.0",
    "croniter>=2.0.0",
    "dingtalk-stream>=0.4.0",
    "python-telegram-bot[socks]>=21.0",
    "lark-oapi>=1.0.0",
    "socksio>=1.0.0",
    "python-socketio>=5.11.0",
    "msgpack>=1.0.8",
    "slack-sdk>=3.26.0",
    "qq-botpy>=1.0.0",
    "python-socks[asyncio]>=2.4.0",
    "prompt-toolkit>=3.0.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
    "requests>=2.28.0",
    "av>=12.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
nanobot = "nanobot.cli.commands:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nanobot"]

[tool.hatch.build.targets.wheel.sources]
"nanobot" = "nanobot"

# Include non-Python files in skills
[tool.hatch.build]
include = [
    "nanobot/**/*.py",
    "nanobot/skills/**/*.md",
    "nanobot/skills/**/*.sh",
]

[tool.hatch.build.targets.sdist]
include = [
    "nanobot/",
    "bridge/",
    "README.md",
    "LICENSE",
]

[tool.hatch.build.targets.wheel.force-include]
"bridge" = "nanobot/bridge"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
